Move PassedInitialConfig interface to its own file.
Create passed-config folder.

Move PASSED_CONFIG InjectionToken to passed-config folder.

Export PassedInitialConfig interface from Public API file since it's a public interface.
We can can now import it like this:
`import { PassedInitialConfig } from 'angular-auth-oidc-client'`
instead of
`import { PassedInitialConfig } from 'angular-auth-oidc-client/lib/auth.module'`
