+++ projects/angular-auth-oidc-client/src/lib/angular-auth-oidc-client.ts
@@ -10,6 +10,8 @@ export * from './config/auth-well-known/auth-well-known-endpoints';
 export * from './config/config.service';
 export * from './config/loader/config-loader';
 export * from './config/openid-configuration';
+export * from './passed-config/passed-config';
+export * from './passed-config/passed-initial-config';
 export * from './interceptor/auth.interceptor';
 export * from './logging/log-level';
 export * from './logging/logger.service';
