+++ projects/angular-auth-oidc-client/src/lib/passed-config/passed-config.ts
@@ -1,0 +1,4 @@
+import { InjectionToken } from '@angular/core';
+import { PassedInitialConfig } from './passed-initial-config';
+
+export const PASSED_CONFIG = new InjectionToken<PassedInitialConfig>('PASSED_CONFIG');
