+++ projects/angular-auth-oidc-client/src/lib/passed-config/passed-initial-config.ts
@@ -1,0 +1,8 @@
+import { Provider } from '@angular/core';
+import { OpenIdConfiguration } from '../config/openid-configuration';
+
+export interface PassedInitialConfig {
+  config?: OpenIdConfiguration | OpenIdConfiguration[];
+  loader?: Provider;
+  storage?: any;
+}
